+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- `Innmind\Immutable\Attempt::mapError()`
+
 ## 5.17.0 - 2025-08-02
 
 ### Added